engine/requirements.txt
google-generativeai
firebase-admin
